requirements.txt
Django==5.1.5
django-filter==24.3
djangorestframework==3.15.2
